refactor: 데이터 변환 함수 파일 모듈화
